Update with new builtins: constants and functions
Update builtins.txt using wiki_builtins.pl
Manually edit builtins.txt to remove extraneous lines for
    ESTATE_ACCESS_BANNED_AGENT_ADD
    PRIM_MEDIA_PERM_ANYONE
    PRIM_MEDIA_PERM_GROUP
    PRIM_MEDIA_PERM_OWNER
These extraneous lines had not value, and were followed in each
case by a correct line with the value for the same constant.  This
is an artifact of either an error in the wiki or an error in
wiki_builtins.pl.  TODO: someone who knows perl (I don't) should
anlayze this error and attempt a workaround.

Without the manual deletion of these lines, running lslint produced
four erroneous messages of the form:
ERROR:: ( 0, 0): '.......' is an event name, and cannot be used
    as a function name.
with the ....... replaced by each of the four constants above.

The resulting sources were turned into an executable on Ubuntu 12.04
by typing "make", producing an executable in the root directory.  This
was briefly tested successfully.  The objects and binary were not
included in the git repository, in keeping with the structure of the
original git repository.

The files lslint and lslint.exe were NOT updated to keep them consistent.
I do not have the capability of making a new lslint.exe.  I do not have
the capability of testing on either the MAC or Windows.